fix(sablefish-package): Removes dll
Not sure why I had a call for building a dll but that was wrong.
Removing this and updating the structure of the package, i.e.,
properly ignoring non-package files, removing removed function from
tests (one is still purposely failing to remind me), etc.
Adds missing packages to the DESCRIPTION file.