Adds commercial discard data
Process to work up the data is similar to that used in 2021, where
the code to process the data is not known for 2019. The methods used
are not standard. And will be changed in a subsequent commit.

FEATURE - there is a flag in data-raw/data_commercial-discard.R that
is on the FIRST LINE of the file to allow for creating the data frame
for discard information using the 2021 methods. If data_analysis is
2021, then the data-processed data frame will be similar to those
created in 2021.
@chantelwetzel-noaa can you please double check the lines labeled with
your name in the .R file. I am not sure how to calculate the input
sample size for the length composition data. :pray: