Stop filling in missing year of the survey index
The missing year, which was denoted with a negative value
allowed for the expected value to be computed but the fit not
included in the likelihood. We do this with Pacific Hake because
we have a survey every other year. The results were in line with
the trajectory but r4ss did not accommodate the value very well
and the plots were jacked up. This should be fixed. But for now,
I just removed it.