fix(data_env_index): Need to use z-scored data
I was not using z-scored and I was using the Bayesian output.
Now, I am z-scoring, though I am using sigma_r of 1.4 where in 2019 and
2021 they used a value of 1.0 to calculate the z-score and I am not sure
why. Melissa is on vacation until the 7th. Nick doesn't know anything
about this process.

Will need to update the bridging. Nick said to use the old DFA analysis
because it was stable when he redid it for us. The bayesian will be a
sensitivity. 2021 used the Bayesian output but Nick is not sure how he
provided a SE because this is not traditional output from the model.