fix path in run script to location of base models
Not sure of a good way to make sure that this is always up to date
because I cannot put it in internal because then the cache of the
document would not update if the location of the base model changed I
do not think. So, I made a note that if I change it in the run script
that it also needs to be changed in 00a.Rmd. Any ideas are welcome.